@@ -10,19 +10,24 @@ import (
 )
 
 type config struct {
-	CompanyName        string
+	CompanyName string
+
 	PostgresSource     string
 	RedisSessionSource string
-	ImageDirectory     string
-	ImageProxyLink     string
+
+	ImageDirectory string
+	ImageProxyLink string
 
 	MailHost     string
 	MailPort     int
 	MailUser     string
 	MailPassword string
 
+	OauthState string
+
 	GoogleOauthClientID     string
 	GoogleOauthClientSecret string
+	GoogleOauthRedirectURL  string
 
 	TokenSigningKey    string
 	TokenEncryptionKey string
@@ -35,18 +40,26 @@ type config struct {
 func (c config) Validate() error {
 	return vd.ValidateStruct(&c,
 		vd.Field(&c.CompanyName, vd.Required),
+
 		vd.Field(&c.PostgresSource, vd.Required),
 		vd.Field(&c.RedisSessionSource, vd.Required),
+
 		vd.Field(&c.ImageDirectory, vd.Required),
 		vd.Field(&c.ImageProxyLink, vd.Required),
+
 		vd.Field(&c.MailHost, vd.Required),
 		vd.Field(&c.MailPort, vd.Required),
 		vd.Field(&c.MailUser, vd.Required),
 		vd.Field(&c.MailPassword, vd.Required),
+
+		vd.Field(&c.OauthState, vd.Required),
 		vd.Field(&c.GoogleOauthClientID, vd.Required),
 		vd.Field(&c.GoogleOauthClientSecret, vd.Required),
+		vd.Field(&c.GoogleOauthRedirectURL, vd.Required),
+
 		vd.Field(&c.TokenSigningKey, vd.Required),
 		vd.Field(&c.TokenEncryptionKey, vd.Required),
+
 		vd.Field(&c.ServerPort, vd.Required),
 		vd.Field(&c.Environment, vd.Required),
 	)
@@ -60,21 +73,29 @@ func ReadConfig(files ...string) {
 	}
 
 	C = config{
-		CompanyName:             "Golang Corp",
-		PostgresSource:          os.Getenv("POSTGRES_SOURCE"),
-		RedisSessionSource:      os.Getenv("SESSION_DB_SOURCE"),
-		ImageDirectory:          os.Getenv("IMAGE_DIRECTORY"),
-		ImageProxyLink:          os.Getenv("IMGPROXY_LINK"),
-		MailHost:                os.Getenv("MAIL_HOST"),
-		MailPort:                MustParseInt[int](os.Getenv("MAIL_PORT")),
-		MailUser:                os.Getenv("MAIL_USER"),
-		MailPassword:            os.Getenv("MAIL_PASSWORD"),
+		CompanyName: "Golang Corp",
+
+		PostgresSource:     os.Getenv("POSTGRES_SOURCE"),
+		RedisSessionSource: os.Getenv("SESSION_DB_SOURCE"),
+
+		ImageDirectory: os.Getenv("IMAGE_DIRECTORY"),
+		ImageProxyLink: os.Getenv("IMGPROXY_LINK"),
+
+		MailHost:     os.Getenv("MAIL_HOST"),
+		MailPort:     MustParseInt[int](os.Getenv("MAIL_PORT")),
+		MailUser:     os.Getenv("MAIL_USER"),
+		MailPassword: os.Getenv("MAIL_PASSWORD"),
+
+		OauthState:              os.Getenv("OAUTH_STATE"),
 		GoogleOauthClientID:     os.Getenv("GOOGLE_OAUTH_CLIENT_ID"),
 		GoogleOauthClientSecret: os.Getenv("GOOGLE_OAUTH_CLIENT_SECRET"),
-		TokenSigningKey:         os.Getenv("TOKEN_SIGNING_KEY"),
-		TokenEncryptionKey:      os.Getenv("TOKEN_ENCRYPTION_KEY"),
-		ServerPort:              os.Getenv("SERVER_PORT"),
-		Environment:             os.Getenv("ENVIRONMENT"),
+		GoogleOauthRedirectURL:  os.Getenv("GOOGLE_OAUTH_REDIRECT_URL"),
+
+		TokenSigningKey:    os.Getenv("TOKEN_SIGNING_KEY"),
+		TokenEncryptionKey: os.Getenv("TOKEN_ENCRYPTION_KEY"),
+
+		ServerPort:  os.Getenv("SERVER_PORT"),
+		Environment: os.Getenv("ENVIRONMENT"),
 	}
 
 	if err := C.Validate(); err != nil {
