@@ -2,7 +2,6 @@ package cookiemanager
 
 import (
 	"context"
-	"errors"
 	"net/http"
 	"time"
 
@@ -31,16 +30,24 @@ type CookieManager struct {
 
 // Add a cookie into the cookie stack.
 func (c *CookieManager) SetCookie(name string, value string, maxAge time.Duration) {
-	http.SetCookie(c.rw, &http.Cookie{
+
+	cookie := &http.Cookie{
 		Name:     name,
 		Value:    value,
 		MaxAge:   int(maxAge.Seconds()),
+		Expires:  time.Now().Add(maxAge),
 		HttpOnly: c.cfg.HttpOnly,
 		Secure:   c.cfg.Secure,
 		SameSite: c.cfg.SameSite,
 		Path:     c.cfg.Path,
 		Domain:   c.cfg.Domain,
-	})
+	}
+
+	if err := cookie.Valid(); err != nil {
+		panic("invalid cookie " + err.Error())
+	}
+
+	http.SetCookie(c.rw, cookie)
 }
 
 // Remove a cookie by name (will expire the cookie).
@@ -49,6 +56,7 @@ func (c *CookieManager) RemoveCookie(name string) {
 		Name:     name,
 		Value:    "",
 		MaxAge:   -1,
+		Expires:  time.Unix(0, 0),
 		HttpOnly: c.cfg.HttpOnly,
 		Secure:   c.cfg.Secure,
 		SameSite: c.cfg.SameSite,
@@ -71,11 +79,11 @@ func LoadManager(cfg CookieConfig) func(next http.Handler) http.Handler {
 }
 
 // Get the cookie manager from context.
-func For(ctx context.Context) (*CookieManager, error) {
+func For(ctx context.Context) *CookieManager {
 	cookieMgr, ok := ctx.Value(config.COOKIE_MANAGER_KEY).(*CookieManager)
 	if !ok {
-		return nil, errors.New("cookie manager not found in context")
+		panic("cookie manager not found in context")
 	}
 
-	return cookieMgr, nil
+	return cookieMgr
 }
