@@ -1,22 +1,19 @@
 package graphql
 
 import (
-	"context"
-	"errors"
 	"log"
 
-	"github.com/99designs/gqlgen/graphql"
 	"github.com/99designs/gqlgen/graphql/handler"
 	"github.com/99designs/gqlgen/graphql/playground"
 	"github.com/go-chi/chi/v5"
-	validation "github.com/go-ozzo/ozzo-validation/v4"
-	"github.com/vektah/gqlparser/v2/gqlerror"
 
 	"github.com/thanishsid/goserver/config"
 	"github.com/thanishsid/goserver/domain"
 	"github.com/thanishsid/goserver/graphql/cookiemanager"
 	"github.com/thanishsid/goserver/graphql/dataloader"
+	"github.com/thanishsid/goserver/graphql/directive"
 	"github.com/thanishsid/goserver/graphql/generated"
+	"github.com/thanishsid/goserver/graphql/middleware"
 	"github.com/thanishsid/goserver/graphql/resolver"
 )
 
@@ -24,35 +21,57 @@ type ServerDeps struct {
 	UserService    domain.UserService
 	ImageService   domain.ImageService
 	SessionService domain.SessionService
+	AuthService    domain.AuthService
 }
 
 func NewServer(deps ServerDeps) *chi.Mux {
 
 	// Create new mux router.
-	mux := chi.NewRouter()
+	r := chi.NewRouter()
 
+	// Cookie Manager Configs.
 	cookieConfig := cookiemanager.CookieConfig{
 		HttpOnly: true,
 		Secure:   config.C.Environment != "development",
 		Path:     "/",
-		Domain:   "127.0.0.1",
 	}
 
 	// Load middleware.
-	mux.Use(
+	r.Use(
+		middleware.LoadRequestMetadataMiddleware(),
 		cookiemanager.LoadManager(cookieConfig),
-		LoadSessionMiddleware(deps.SessionService),
+		middleware.LoadSessionMiddleware(deps.SessionService),
 		dataloader.Middleware(dataloader.NewDataloader(deps.UserService, deps.ImageService)),
 	)
 
+	// Mount the Oauth2 login http routes.
+	r.Mount("/auth", LoadOauthRoutes(deps.AuthService))
+
+	// Load graphql handler onto http mux in the /query endpoint
+	r.Handle("/query", NewGraphqlHandler(deps))
+
+	// Will load GraphiQL if in Development Environment
+	if config.C.Environment == "development" {
+		r.Handle("/", playground.Handler("GraphQL playground", "/query"))
+		log.Printf("connect to http://localhost:%s/ for GraphQL playground", config.C.ServerPort)
+	}
+
+	return r
+}
+
+// Create a new graphql handler with configs.
+func NewGraphqlHandler(deps ServerDeps) *handler.Server {
 	// Graphql configs.
 	conf := generated.Config{Resolvers: &resolver.Resolver{
 		UserService:    deps.UserService,
 		ImageService:   deps.ImageService,
 		SessionService: deps.SessionService,
+		AuthService:    deps.AuthService,
 	},
 		Directives: generated.DirectiveRoot{
-			EnforceAction: enforceActionDirective,
+			Authenticated:   directive.Authenticated,
+			Unauthenticated: directive.Unauthenticated,
+			Authorize:       directive.Authorize,
 		},
 	}
 
@@ -62,26 +81,5 @@ func NewServer(deps ServerDeps) *chi.Mux {
 	// Graphql Error handler.
 	gqlHandler.SetErrorPresenter(gqlErrFunc)
 
-	mux.Handle("/", playground.Handler("GraphQL playground", "/query"))
-	mux.Handle("/query", gqlHandler)
-
-	log.Printf("connect to http://localhost:%s/ for GraphQL playground", config.C.ServerPort)
-
-	return mux
-}
-
-func gqlErrFunc(ctx context.Context, err error) *gqlerror.Error {
-	gqlErr := graphql.DefaultErrorPresenter(ctx, err)
-
-	if gqlErr.Extensions == nil {
-		gqlErr.Extensions = make(map[string]interface{})
-	}
-
-	var vdErr validation.Errors
-	if errors.As(err, &vdErr) {
-		gqlErr.Message = "validation error"
-		gqlErr.Extensions["validationErrors"] = vdErr
-	}
-
-	return gqlErr
+	return gqlHandler
 }
