@@ -177,7 +177,7 @@ type GetUserByEmailRow struct {
 	Email        string
 	FullName     string
 	Role         string
-	PasswordHash string
+	PasswordHash null.String
 	PictureID    uuid.NullUUID
 	CreatedAt    time.Time
 	UpdatedAt    time.Time
@@ -226,7 +226,7 @@ type GetUserByIdRow struct {
 	Email        string
 	FullName     string
 	Role         string
-	PasswordHash string
+	PasswordHash null.String
 	PictureID    uuid.NullUUID
 	CreatedAt    time.Time
 	UpdatedAt    time.Time
@@ -299,7 +299,7 @@ type InsertOrUpdateUserParams struct {
 	Email        string
 	FullName     string
 	Role         string
-	PasswordHash string
+	PasswordHash null.String
 	PictureID    uuid.NullUUID
 	CreatedAt    time.Time
 	UpdatedAt    time.Time
